update version 02/12/2017

+++ publishable/database/migrations/2016_01_01_000000_add_voyager_user_fields.php
@@ -11,7 +11,7 @@ class AddVoyagerUserFields extends Migration
     {
         Schema::table('users', function ($table) {
             if (!Schema::hasColumn('users', 'avatar')) {
-                $table->string('avatar')->nullable()->after('email');
+                $table->string('avatar')->nullable()->after('email')->default('users/default.png');
             }
             $table->integer('role_id')->nullable()->after('id');
         });
@@ -22,9 +22,15 @@ class AddVoyagerUserFields extends Migration
      */
     public function down()
     {
-        Schema::table('users', function ($table) {
-            $table->dropColumn('avatar');
-            $table->dropColumn('role_id');
-        });
+        if (Schema::hasColumn('users', 'avatar')) {
+            Schema::table('users', function ($table) {
+                $table->dropColumn('avatar');
+            });
+        }
+        if (Schema::hasColumn('users', 'role_id')) {
+            Schema::table('users', function ($table) {
+                $table->dropColumn('role_id');
+            });
+        }
     }
 }
